API connections made
