+++ services/ApiService.ts
@@ -1,6 +1,6 @@
 import axios from "axios"
 
-const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJiYXJpc3llbmlndW4iLCJzY29wZXMiOlsidXNlcjpyZWFkIiwidXNlcjp3cml0ZSJdLCJleHAiOjE3MDU5NTMxMTd9.B2COfGzCoXd7Wag5Ql_SY_p5DvFqczZp7axBrcBJYAo";
+const token = "YOUR_JWT_KEY";
 
 /*
 Examples 
@@ -17,15 +17,15 @@ const getAvailableUsers = async (isAvailable: string) => {
 */
 
 export const setOtpChoice = async (otpIdentifier: string) => {
-    return axios.post("/auth/otp/set", {otp_type: "Email", otp_identifier: otpIdentifier}, {headers: {Authorization: `Bearer ${token}`}});
+    return axios.post("https://platform-omggm8hk.b4a.run/auth/otp/set", {otp_type: "Email", otp_identifier: otpIdentifier}, {headers: {Authorization: `Bearer ${token}`}});
 }
 
 export const sendPasswordResetRequest = () => {
-    axios.post("/auth/password-reset/request", {}, {headers: {Authorization: `Bearer ${token}`}});
+    axios.post("https://platform-omggm8hk.b4a.run/auth/password-reset/request", {}, {headers: {Authorization: `Bearer ${token}`}});
 }
 
 export const verifyOtpAndUpdatePassword = (otpCode: string, newPassword: string) => {
-    return axios.post(`/auth/password-reset/verify?otp_code=${otpCode}&new_password=${newPassword}`, {}, {headers: {Authorization: `Bearer ${token}`}});
+    return axios.post(`https://platform-omggm8hk.b4a.run/auth/password-reset/verify?otp_code=${otpCode}&new_password=${newPassword}`, {}, {headers: {Authorization: `Bearer ${token}`}});
 }
